rename: SubmittedForms 하위 컴포넌트 파일 정리
